Refactor to put waterfall and trace behind the same observability tab
