+++ frontend/app/src/pages/main/v1/workflow-runs-v1/$run/v2components/step-run-detail/observability/otel-span-adapter.ts
